+++ src/lib/Scenes/City/Components/AllEvents.tsx
@@ -3,6 +3,7 @@ import { EventSection } from "lib/Scenes/City/Components/EventSection"
 import { BucketKey, BucketResults } from "lib/Scenes/Map/Bucket"
 import React from "react"
 import { FlatList, ViewProperties } from "react-native"
+import { RelayProp } from "react-relay"
 import { BMWEventSection } from "./BMWEventSection"
 import { FairEventSection } from "./FairEventSection"
 import { SavedEventSection } from "./SavedEventSection"
@@ -12,6 +13,7 @@ interface Props extends ViewProperties {
   buckets: BucketResults
   cityName: string
   sponsoredContent: { introText: string; artGuideUrl: string }
+  relay: RelayProp
 }
 
 interface State {
@@ -104,20 +106,28 @@ export class AllEvents extends React.Component<Props, State> {
   }
 
   renderItem = ({ item: { data, type } }) => {
-    const { sponsoredContent } = this.props
+    const { sponsoredContent, cityName } = this.props
     switch (type) {
       case "fairs":
         return <FairEventSection data={data} />
       case "galleries":
-        return <EventSection title="Gallery shows" data={data} />
+        return <EventSection title="Gallery shows" data={data} relay={this.props.relay} />
       case "museums":
-        return <EventSection title="Museum shows" data={data} />
+        return <EventSection title="Museum shows" data={data} relay={this.props.relay} />
       case "opening":
-        return <EventSection title="Opening shows" data={data} />
+        return <EventSection title="Opening shows" data={data} relay={this.props.relay} />
       case "closing":
-        return <EventSection title="Closing shows" data={data} />
+        return <EventSection title="Closing shows" data={data} relay={this.props.relay} />
       case "bmw":
-        return <BMWEventSection title="BMW Art Guide" sponsoredContent={sponsoredContent} data={data} />
+        return (
+          <BMWEventSection
+            title="BMW Art Guide"
+            sponsoredContent={sponsoredContent}
+            data={data}
+            cityName={cityName}
+            relay={this.props.relay}
+          />
+        )
       case "saved":
         return <SavedEventSection data={data} />
       case "header":